removing
removing this list because people are bothered. Some accounts are definitely not malicious as the whole point is to be able to eventually classify repo content. But as its unclear from dataset which users are "malicious" and "benign" appearing on list is upsetting to some.

This was an unintended consequence. Though, these people are likely already on lists... they just weren't open source